requirements.txt
pylint
autopep8
typeguard
cmake
ninja
pytest
pytest-dependency
pytest-html-reporter
typer
pyaml
tensorflow==2.5.1
tensorflow_probability==0.12.2
numpy
scikit-learn==0.24.0
matplotlib
tqdm
pillow
librosa
netron
GPUtil
bincopy
pyserial
patool==1.12
opencv-python
git+https://stash.silabs.com/scm/iot_dsa/tflite_model_parameters.git
